+++ tx-backend/src/test/java/org/eclipse/tractusx/traceability/integration/importdata/ImportControllerIT.java
@@ -478,5 +478,4 @@ class ImportControllerIT extends IntegrationTestSpecification {
 
     }
 
-
 }