+++ tx-models/src/main/java/assets/importpoc/ImportJobStatusResponse.java
@@ -18,6 +18,9 @@
  ********************************************************************************/
 package assets.importpoc;
 
+import io.swagger.annotations.ApiModel;
+
+@ApiModel
 public enum ImportJobStatusResponse {
     INITIALIZING, RUNNING, CANCELLED, COMPLETED
 }